// projects/learners-point-app/src/app/shared/constants/enums/xml-headers.ts
export enum XmlConstants {

  XSD_SCHEMA_GENERATOR = 'XSD Schema Generator',
  GENERATE_XSD = 'Generate XSD',
  XSD_GENERATOR = "XSD Generator",
  CONVERT_XML_TO_JSON = 'Convert XML to JSON',
  CONVERT_JSON_TO_XML = 'Convert JSON to XML',
  XML_TO_JSON_CONVERTER = 'XML to JSON Converter',
  JSON_TO_XML_CONVERTER = 'JSON to XML Converter',
  CONVERT_JSON_TO_YAML = "Convert JSON to YAML",
  JSON_TO_YAML_CONVERTER = "JSON to YAML Converter",
  YAML_TO_JSON_CONVERTER = "YAML to JSON Converter",
  CONVERT_YAML_TO_JSON = "Convert YAML to JSON",
  CSV_TO_XML_CONVERTER = "CSV to XML Converter",
  CONVERT_CSV_FILE_TO_XML_FILE = 'Convert CSV file to XML file',
  CSV_TO_JSON_CONVERTER = 'CSV to JSON Converter',
  CONVERT_CSV_TO_JSON = 'Convert CSV to JSON',
  XML_FORMATTER = 'XML Formatter',
  FORMAT_XML = 'Format XML',
  FORMAT_JSON = 'Format JSON',
  JSON_FORMATTER = 'JSON Formatter',
  HTML_FORMATTER = 'HTML Formatter',
  FORMAT_HTML = 'Format HTML',
  SQL_FORMATTER = 'SQL Formatter',
  FORMAT_SQL = 'Format SQL',
  GENERATE_JAVA_CLASS = 'Generate java class',
  XML_TO_JAVA_CLASS = 'XML to Java class',
  JSON_TO_JAVA_CLASS = 'JSON to Java Class'



}